PMM-7873 move common types
