+++ public/app/percona/backup/Backup.types.ts
@@ -4,3 +4,18 @@ export enum TabKeys {
   locations = 'locations',
   restore = 'restore',
 }
+
+export enum DataModel {
+  DATA_MODEL_INVALID = 'DATA_MODEL_INVALID',
+  PHYSICAL = 'PHYSICAL',
+  LOGICAL = 'LOGICAL',
+}
+
+export enum Status {
+  STATUS_INVALID = 'STATUS_INVALID',
+  PENDING = 'PENDING',
+  IN_PROGRESS = 'IN_PROGRESS',
+  PAUSED = 'PAUSED',
+  SUCCESS = 'SUCCESS',
+  ERROR = 'ERROR',
+}
