+++ public/app/percona/backup/components/Status/Status.tsx
@@ -1,6 +1,6 @@
 import React, { FC } from 'react';
 import { useTheme } from '@grafana/ui';
-import { formatStatus } from '../BackupInventory.utils';
+import { formatStatus } from '../BackupInventory/BackupInventory.utils';
 import { StatusProps } from './Status.types';
 import { getStyles } from './Status.styles';
